acme_project/requirements.txt
asgiref==3.7.2
Django==3.2.16
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
Pillow==9.3.0
pytz==2023.3.post1
sqlparse==0.4.4
